+++ cs3560_grading_lib/common.py
@@ -16,7 +16,8 @@ NAME_POOL = ["herta", "cat", "dog", "dolphin", "falcon", "dandilion", "fox", "je
 
 
 def is_debug_mode(
-    variable_name: str = "DEBUG", vals_for_true: list[str] = ["true", "t", "on", "1"]
+    variable_name: str = "DEBUG",
+    vals_for_true: list[str] = ["true", "t", "on", "1"],
 ) -> bool:
     """Return True if the DEBUG envrionment variable is presence with value representing 'True'."""
     raw_val = os.environ.get(variable_name, None)