+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - Fix `run_executable` not capture stdout with pipe.
+- Fix parsing problem with Makefile where there is no empty line
+  between the rules.
 
 ## v0.0.3